+++ test/web/10-api.js
@@ -7,7 +7,6 @@ import sinon from 'sinon';
 import {mockData} from './mock.data.js';
 
 const ACCOUNT_ID = 'urn:uuid:ffaf5d84-7dc2-4f7b-9825-cc8d2e5a5d06';
-const KMS_MODULE = 'ssm-v1';
 const KMS_BASE_URL = `${window.location.origin}/kms`;
 const EDV_BASE_URL = `${window.location.origin}/edvs`;
 
@@ -15,7 +14,6 @@ describe('Profile Manager API', () => {
   describe('createProfile API', () => {
     it('successfully creates a profile', async () => {
       const profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -60,7 +58,6 @@ describe('Profile Manager API', () => {
           didOptions
         }).returns({id: profileDid});
         const profileManager = new ProfileManager({
-          kmsModule: KMS_MODULE,
           kmsBaseUrl: KMS_BASE_URL,
           edvBaseUrl: EDV_BASE_URL,
           profileService
@@ -98,7 +95,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -162,7 +158,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -225,7 +220,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -335,7 +329,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -435,7 +428,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -501,7 +493,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -597,7 +588,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -672,7 +662,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(async () => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -808,7 +797,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(() => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -852,7 +840,6 @@ describe('Profile Manager API', () => {
     let profileManager;
     beforeEach(() => {
       profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL
       });
@@ -896,7 +883,6 @@ describe('Profile Manager API', () => {
       let error;
       let result;
       const profileManager = new ProfileManager({
-        kmsModule: KMS_MODULE,
         kmsBaseUrl: KMS_BASE_URL,
         edvBaseUrl: EDV_BASE_URL,
         // intentionally make zcap expired