+++ test/web/20-collection.js
@@ -5,7 +5,6 @@ import {ProfileManager} from 'bedrock-web-profile-manager';
 import {EdvClient} from '@digitalbazaar/edv-client';
 
 const ACCOUNT_ID = 'urn:uuid:ffaf5d84-7dc2-4f7b-9825-cc8d2e5a5d06';
-const KMS_MODULE = 'ssm-v1';
 const KMS_BASE_URL = `${window.location.origin}/kms`;
 const EDV_BASE_URL = `${window.location.origin}/edvs`;
 
@@ -13,7 +12,6 @@ describe('Collection API', () => {
   let profileManager;
   beforeEach(async () => {
     profileManager = new ProfileManager({
-      kmsModule: KMS_MODULE,
       kmsBaseUrl: KMS_BASE_URL,
       edvBaseUrl: EDV_BASE_URL
     });